+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 3.11.4 Release (24.09.2018)
 * Fixed rewards weren't dispatched properly while using %PLAYER% placeholder (it was always executed by players)
+* Fixed endwave commands not working
 
 ### 3.11.3 Release (07/23.09.2018)
 * Dropped 1.9-1.10 support